+++ Scripts/Items/Artifacts/Consumables/VialofArmorEssence.cs
@@ -123,7 +123,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)1); // version
+            writer.Write(1); // version
         }
 
         public override void Deserialize(GenericReader reader)
