+++ Scripts/Items/Books/SpecialScrollBooks/RecipeBook/RecipeFilter.cs
@@ -42,15 +42,15 @@ namespace Server.Items
         {
             if (IsDefault)
             {
-                writer.Write((int)0);
+                writer.Write(0);
             }
             else
             {
                 writer.Write(1);
 
-                writer.Write((int)Skill);
-                writer.Write((int)Expansion);
-                writer.Write((int)Amount);
+                writer.Write(Skill);
+                writer.Write(Expansion);
+                writer.Write(Amount);
             }
         }
     }
